moved more functions to poutine.util
